pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-cop30"
version = "0.1.0"
description = "Agentic AI orchestration demo"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  # Core
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "pydantic>=2.7",

  # LLMs
  "openai>=1.54",

  # Web / API
  "fastapi>=0.115",
  "uvicorn>=0.30",

  # Text parsing / scraping
  "beautifulsoup4>=4.12",
  "lxml>=5.3",
  "rapidfuzz>=3.9",

  # Utilities
  "tenacity>=8.2",
  "loguru>=0.7",
  "rich>=13.7",

  # Optional but recommended for network stability
  "brotli>=1.1",
  "certifi",
  "charset-normalizer",

  # Memory / storage
  "sqlite-utils>=3.36",

  # Scheduling / async orchestration
  "anyio>=4.0",

  # Graph orchestration
  "langgraph>=0.2",

  # JSON perf
  "orjson>=3.10",

  # Windows SSL fixes
  "idna>=3.7",
]

[tool.setuptools]
packages = ["agents", "topics"]

[tool.setuptools.package-data]
agents = ["**/*.txt", "**/*.json", "**/*.md"]
topics = ["**/*.txt", "**/*.json", "**/*.md"]
